move get_value_3d logic to scalar field with small changes so image works
